Refactor the name of package.
