+++ zipkin-server-2/src/main/java/com/example/zipkinserver_2/ZipkinServer2Application.java
@@ -1,4 +1,4 @@
-package com.example.zipkinserver2;
+package com.example.zipkinserver_2;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
