+++ zipkin-server-2/src/main/java/com/example/zipkinserver_2/business/src/service/Impl/Zipkin2ServiceImpl.java
@@ -1,6 +1,6 @@
-package com.example.zipkinserver2.business.src.service.Impl;
+package com.example.zipkinserver_2.business.src.service.Impl;
 
-import com.example.zipkinserver2.business.src.service.Zipkin2Service;
+import com.example.zipkinserver_2.business.src.service.Zipkin2Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.ParameterizedTypeReference;
