+++ zipkin-server-2/src/main/java/com/example/zipkinserver_2/business/src/service/Zipkin2Service.java
@@ -1,4 +1,4 @@
-package com.example.zipkinserver2.business.src.service;
+package com.example.zipkinserver_2.business.src.service;
 
 public interface Zipkin2Service {
